Updating PathPlanning for RR1.0
