+++ PathPlanning/build.gradle
@@ -14,4 +14,5 @@ repositories {
 
 dependencies {
     implementation 'com.github.rh-robotics:MeepMeep:v1.0.0'
+    implementation 'com.acmerobotics.roadrunner:MeepMeep:0.1.6'
 }
